refactor(data): wave-5 slice C embed rename and 400-line ratchets (#4292)
Rename tech-effect embed to *.gen.dart, split colonial victory-config
params and civilian-build scoring for the 400 physical-line lib ceiling,
and ratchet repo.data_lib_file_size / repo.data_test_file_size from 500
to 400 with SPEC and checker test updates.
