+++ SPEC/ui/tech-tree-widget.md
@@ -93,7 +93,7 @@ States are mutually exclusive; “in progress” takes precedence over “availa
 ## Data
 
 - **Catalog:** `techCatalog` from colonizethis_data (id, era, category, cost, prerequisiteIds, regimentUnlockIds, shipUnlockIds, etc.). Display name: humanized from id (e.g. `road_construction` → “Road Construction”) unless a display-name map is added.
-- **Effect summary (dialog):** Authored per-tech lines live in `packages/colonizethis_data/lib/src/data/tech_effect_summary.yaml` (embedded at runtime via `tech_effect_summary_embed.dart`). Each line has a stable id (`techEffectSummary_<techId>_<index>`) and English template text. The Flutter app resolves those ids through **`AppLocalizations`** (keys merged into `app/lib/l10n/arb/app_en.arb` by `tool/generate_tech_effect_l10n.dart`). Regiment/ship unlock bullets and category fallback strings are separate ARB messages. Do not add large per-tech `switch`es in the widget for this copy. **Shared helper:** `buildTechEffectSummaryLines` in `app/lib/features/game/widgets/technology/tech_effect_summary.dart` is the single source for effect-line strings; the Tree description dialog and Choose-tech rows (default 1–2 lines + Details) both call it (Refs #4222).
+- **Effect summary (dialog):** Authored per-tech lines live in `packages/colonizethis_data/lib/src/data/tech_effect_summary.yaml` (embedded at runtime via `tech_effect_summary_embed.gen.dart`). Each line has a stable id (`techEffectSummary_<techId>_<index>`) and English template text. The Flutter app resolves those ids through **`AppLocalizations`** (keys merged into `app/lib/l10n/arb/app_en.arb` by `tool/generate_tech_effect_l10n.dart`). Regiment/ship unlock bullets and category fallback strings are separate ARB messages. Do not add large per-tech `switch`es in the widget for this copy. **Shared helper:** `buildTechEffectSummaryLines` in `app/lib/features/game/widgets/technology/tech_effect_summary.dart` is the single source for effect-line strings; the Tree description dialog and Choose-tech rows (default 1–2 lines + Details) both call it (Refs #4222).
 - **Player:** `Player.techUnlocked`, `Player.researchProgressByTechId`, `Player.researchSlots`. **Researchable** set is required: techs whose prerequisites are all in `techUnlocked`, tech not in `techUnlocked`, and not already in `researchProgressByTechId`.
 
 ## Acceptance criteria
