+++ packages/colonizethis_data/lib/src/tech_effect_summary_embed.gen.dart
