+++ tool/ct_repo_lint_manifest.yaml
@@ -1317,7 +1317,7 @@ rules:
 
   - rule_id: repo.data_lib_file_size
     group: structure
-    title: colonizethis_data lib/src files must stay at or below 500 non-comment lines (Refs #4072)
+    title: colonizethis_data lib/src files must stay at or below 400 physical lines (Refs #4292)
     spec: SPEC/program/dart-file-non-comment-line-size.md
     runner: dart
     script: tool/check_data_lib_file_size.dart
@@ -1332,7 +1332,7 @@ rules:
 
   - rule_id: repo.data_test_file_size
     group: testing
-    title: colonizethis_data test files must stay at or below 500 physical lines (Refs #4121)
+    title: colonizethis_data test files must stay at or below 400 physical lines (Refs #4292)
     spec: SPEC/program/data-test-file-size.md
     runner: dart
     script: tool/check_data_test_file_size.dart
